feat: separate url params in chasqui topic
Move messageToUri parsing inside chasquiCompress function
Add test for message format in qr.chasquiCompress

All url parameters in a sufficiently large message given to chasquiCompress will now be encoded as key-value pairs in the JSON payload that gets posted to chasqui.  This simplifies parsing on the mobile side after receiving a chasqui url as a message.